pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-qa-framework"
version = "1.0.0"
description = "Autonomous AI-Driven Website QA Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.40.0",
    "anthropic>=0.39.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "Pillow>=10.0.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
qa-framework = "src.cli:cli"

[tool.setuptools.packages.find]
include = ["src*"]
